权限

+++ metaq-service-clients/metaq-uaa-service-client/src/main/java/cn/metaq/uaa/dto/ClientDetailsDTO.java
@@ -32,9 +32,9 @@ public class ClientDetailsDTO implements IDto{
 
 	private String autoApproveScopes;
 
-	private Long accessTokenValiditySeconds;
+	private Integer accessTokenValiditySeconds;
 
-	private Long refreshTokenValiditySeconds;
+	private Integer refreshTokenValiditySeconds;
 
 	private String additionalInformation;
